+++ src/core/Vector3.java
@@ -6,11 +6,11 @@ public class Vector3 extends Scalar3{
 		super();
 	}
 	
-	public Vector3(float x, float y, float z) {
+	public Vector3(double x, double y, double z) {
 		super(x, y, z);
 	}
 
-	public Vector3(float[] v) {
+	public Vector3(double[] v) {
 		super(v);
 	}
 
@@ -26,13 +26,13 @@ public class Vector3 extends Scalar3{
 		return new Vector3(v1.getX() + v2.getX(), v1.getY() + v2.getY(), v1.getZ() + v2.getZ());
 	}
 
-	public static float dot(Vector3 v1, Vector3 v2) {
+	public static double dot(Vector3 v1, Vector3 v2) {
 		return v1.getX()*v2.getX()+v1.getY()*v2.getY()+v1.getZ()*v2.getZ();
 	}
 
 	public static Vector3 cross(Vector3 v1, Vector3 v2) {
-		float x1 = v1.getX(), y1 = v1.getY(), z1 = v1.getZ();
-		float x2 = v2.getX(), y2 = v2.getY(), z2 = v2.getZ();
+		double x1 = v1.getX(), y1 = v1.getY(), z1 = v1.getZ();
+		double x2 = v2.getX(), y2 = v2.getY(), z2 = v2.getZ();
 		return new Vector3(y1*z2-y2*z1, z1*x2-z2*x1, x1*y2-x2*y1);
 	}
 
@@ -40,8 +40,8 @@ public class Vector3 extends Scalar3{
 		return (colinear_scale(v1, v2) != 0);
 	}
 
-	public static float colinear_scale(Vector3 v1, Vector3 v2) {
-		float scale;
+	public static double colinear_scale(Vector3 v1, Vector3 v2) {
+		double scale;
 		if (v1.getX() != 0 && v2.getX() != 0)
 			scale = v1.getX()/v2.getX();
 		else if (v1.getY() != 0 && v2.getY() != 0)
@@ -53,18 +53,18 @@ public class Vector3 extends Scalar3{
 		return (v1.equals(v2.scaled(scale))?scale:0);
 	}
 
-	public Vector3 scaled(float s) {
+	public Vector3 scaled(double s) {
 		return new Vector3(x*s, y*s, z*s);
 	}
 	
-	public void scale(float s) {
+	public void scale(double s) {
 		this.x *= s;
 		this.y *= s;
 		this.z *= s;
 	}
 
 	public Vector3 product(Matrix3x3 m) {
-		float[][] mData = m.getData();
+		double[][] mData = m.getData();
 		return new Vector3(
 				x*mData[0][0] + y*mData[1][0] + z*mData[2][0],
 				x*mData[0][1] + y*mData[1][1] + z*mData[2][1],
@@ -72,23 +72,23 @@ public class Vector3 extends Scalar3{
 	}
 
 	public Vector3 product(Matrix3x4 m) {
-		float[][] mData = m.getData();
+		double[][] mData = m.getData();
 		return new Vector3(
 				x*mData[0][0] + y*mData[1][0] + z*mData[2][0] + mData[3][0],
 				x*mData[0][1] + y*mData[1][1] + z*mData[2][1] + mData[3][1],
 				x*mData[0][2] + y*mData[1][2] + z*mData[2][2] + mData[3][2]);
 	}
 	
-	public float getBlockMagnitude() {
+	public double getBlockMagnitude() {
 		return Math.abs(x)+Math.abs(y)+Math.abs(z);
 	}
 
-	public float getSquareMagnitude() {
+	public double getSquareMagnitude() {
 		return x*x+y*y+z*z;
 	}
 	
-	public float getEuclideanMagnitude() {
-		return (float) Math.sqrt(getSquareMagnitude());
+	public double getEuclideanMagnitude() {
+		return Math.sqrt(getSquareMagnitude());
 	}
 
 	public void normalize() {
@@ -96,7 +96,7 @@ public class Vector3 extends Scalar3{
 	}
 
     public Vector3 getNormalized() {
-        float scale = 1/ getEuclideanMagnitude();
+        double scale = 1/ getEuclideanMagnitude();
         return new Vector3(getX()*scale, getY()*scale, getZ()*scale);
     }
